Move VPP datastore methods to vpp-specific file
