+++ server/datastore/mysql/apple_mdm.go
@@ -7,7 +7,6 @@ import (
 	"crypto/cipher"
 	"crypto/rand"
 	"database/sql"
-	"encoding/base64"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
@@ -4753,127 +4752,6 @@ WHERE
 	return ctxerr.Wrap(ctx, err, "updating abm_token")
 }
 
-func (ds *Datastore) InsertVPPToken(ctx context.Context, tok *fleet.VPPTokenData, teamID *uint, nullTeam fleet.NullTeamType) (*fleet.VPPTokenDB, error) {
-	insertStmt := `
-	INSERT INTO
-		vpp_tokens (
-			organization_name,
-			location,
-			renew_at,
-			token,
-			team_id,
-			null_team_type
-		)
-	VALUES (?, ?, ?, ?, ?, ?)
-`
-
-	if teamID != nil && nullTeam != fleet.NullTeamNone {
-		return nil, ctxerr.Errorf(ctx, "nullTeam must be set to NullTeamNone if teamID is present")
-	}
-
-	if err := ds.checkVPPNullTeam(ctx, nil, nullTeam); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "checking vpp token null team")
-	}
-
-	tokRawBytes, err := base64.StdEncoding.DecodeString(tok.Token)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "decoding raw vpp token data")
-	}
-
-	var tokRaw fleet.VPPTokenRaw
-	if err := json.Unmarshal(tokRawBytes, &tokRaw); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "unmarshalling raw vpp token")
-	}
-
-	exp, err := time.Parse(fleet.VPPTimeFormat, tokRaw.ExpDate)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "parsing vpp token expiration date")
-	}
-	exp = exp.UTC()
-
-	tokEnc, err := encrypt([]byte(tok.Token), ds.serverPrivateKey)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "encrypt token with datastore.serverPrivateKey")
-	}
-
-	vppTokenDB := &fleet.VPPTokenDB{
-		OrgName:   tokRaw.OrgName,
-		Location:  tok.Location,
-		RenewDate: exp,
-		Token:     tok.Token,
-		TeamID:    teamID,
-		NullTeam:  nullTeam,
-	}
-
-	res, err := ds.writer(ctx).ExecContext(
-		ctx,
-		insertStmt,
-		vppTokenDB.OrgName,
-		vppTokenDB.Location,
-		exp,
-		tokEnc,
-		vppTokenDB.TeamID,
-		vppTokenDB.NullTeam,
-	)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "inserting vpp token")
-	}
-
-	id, _ := res.LastInsertId()
-
-	vppTokenDB.ID = uint(id)
-
-	return vppTokenDB, nil
-}
-
-func (ds *Datastore) GetVPPToken(ctx context.Context, tokenID uint) (*fleet.VPPTokenDB, error) {
-	stmt := `
-	SELECT
-		id,
-		organization_name,
-		location,
-		renew_at,
-		token,
-		team_id,
-		null_team_type
-	FROM
-		vpp_tokens
-	WHERE
-		id = ?
-`
-
-	var tokEnc struct {
-		ID        uint               `db:"id"`
-		OrgName   string             `db:"organization_name"`
-		Location  string             `db:"location"`
-		RenewDate time.Time          `db:"renew_at"`
-		Token     []byte             `db:"token"`
-		TeamID    *uint              `db:"team_id"`
-		NullTeam  fleet.NullTeamType `db:"null_team_type"`
-	}
-
-	if err := sqlx.GetContext(ctx, ds.reader(ctx), &tokEnc, stmt, tokenID); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "selecting vpp token from db")
-	}
-
-	tokDec, err := decrypt(tokEnc.Token, ds.serverPrivateKey)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "decrypting vpp token with serverPrivateKey")
-	}
-
-	tok := &fleet.VPPTokenDB{
-		ID:        tokEnc.ID,
-		OrgName:   tokEnc.OrgName,
-		Location:  tokEnc.Location,
-		RenewDate: tokEnc.RenewDate,
-		Token:     string(tokDec),
-		TeamID:    tokEnc.TeamID,
-		NullTeam:  tokEnc.NullTeam,
-	}
-
-	return tok, nil
-}
-
 func (ds *Datastore) InsertABMToken(ctx context.Context, tok *fleet.ABMToken) (*fleet.ABMToken, error) {
 	const stmt = `
 INSERT INTO
@@ -4919,233 +4797,6 @@ VALUES (?, ?, ?, ?, ?, ?, ?, ?)
 	tok.MDMServerURL = url
 
 	return tok, nil
-}
-
-func (ds *Datastore) UpdateVPPToken(ctx context.Context, tok *fleet.VPPTokenDB) error {
-	stmt := `
-	UPDATE
-		vpp_tokens
-	SET
-		organization_name = ?,
-		location = ?,
-		renew_at = ?,
-		token = ?,
-		team_id = ?,
-		null_team_type = ?
-	WHERE
-		id = ?
-`
-
-	if tok.TeamID != nil && tok.NullTeam != fleet.NullTeamNone {
-		return ctxerr.Errorf(ctx, "NullTeam must be set to NullTeamNone if TeamID is present")
-	}
-
-	if err := ds.checkVPPNullTeam(ctx, tok.TeamID, tok.NullTeam); err != nil {
-		return ctxerr.Wrap(ctx, err, "checking vpp null team for update")
-	}
-
-	tokEnc, err := encrypt([]byte(tok.Token), ds.serverPrivateKey)
-	if err != nil {
-		return ctxerr.Wrap(ctx, err, "encrypt token with datastore.serverPrivateKey")
-	}
-
-	_, err = ds.writer(ctx).ExecContext(
-		ctx,
-		stmt,
-		tok.OrgName,
-		tok.Location,
-		tok.RenewDate,
-		tokEnc,
-		tok.TeamID,
-		tok.NullTeam,
-		tok.ID,
-	)
-	if err != nil {
-		return ctxerr.Wrap(ctx, err, "updating vpp token")
-	}
-
-	return nil
-}
-
-func (ds *Datastore) UpdateVPPTokenTeam(ctx context.Context, id uint, teamID *uint, nullTeam fleet.NullTeamType) error {
-	stmt := `
-	UPDATE
-		vpp_tokens
-	SET
-		team_id = ?,
-		null_team_type = ?
-	WHERE
-		id = ?
-`
-	if teamID != nil && nullTeam != fleet.NullTeamNone {
-		return ctxerr.Errorf(ctx, "NullTeam must be set to NullTeamNone if TeamID is present")
-	}
-
-	if err := ds.checkVPPNullTeam(ctx, &id, nullTeam); err != nil {
-		return ctxerr.Wrap(ctx, err, "checking vpp null team for update")
-	}
-
-	if _, err := ds.writer(ctx).ExecContext(ctx, stmt, teamID, nullTeam, id); err != nil {
-		return ctxerr.Wrap(ctx, err, "updating vpp token team")
-	}
-
-	return nil
-}
-
-func (ds *Datastore) DeleteVPPToken(ctx context.Context, tokenID uint) error {
-	_, err := ds.writer(ctx).ExecContext(ctx, `DELETE FROM vpp_tokens WHERE id = ?`, tokenID)
-	if err != nil {
-		return ctxerr.Wrap(ctx, err, "deleting vpp token")
-	}
-
-	return nil
-}
-
-func (ds *Datastore) ListVPPTokens(ctx context.Context) ([]fleet.VPPTokenDB, error) {
-	stmt := `
-	SELECT
-		id,
-		organization_name,
-		location,
-		renew_at,
-		token,
-		team_id,
-		null_team_type
-	FROM
-		vpp_tokens
-`
-	var tokEncs []struct {
-		ID        uint               `db:"id"`
-		OrgName   string             `db:"organization_name"`
-		Location  string             `db:"location"`
-		RenewDate time.Time          `db:"renew_at"`
-		Token     []byte             `db:"token"`
-		TeamID    *uint              `db:"team_id"`
-		NullTeam  fleet.NullTeamType `db:"null_team_type"`
-	}
-
-	if err := sqlx.SelectContext(ctx, ds.reader(ctx), &tokEncs, stmt); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "selecting vpp tokens from db")
-	}
-
-	var tokens []fleet.VPPTokenDB
-
-	for _, tokEnc := range tokEncs {
-		tokDec, err := decrypt(tokEnc.Token, ds.serverPrivateKey)
-		if err != nil {
-			return nil, ctxerr.Wrap(ctx, err, "decrypting vpp token with serverPrivateKey")
-		}
-
-		tokens = append(tokens, fleet.VPPTokenDB{
-			ID:        tokEnc.ID,
-			OrgName:   tokEnc.OrgName,
-			Location:  tokEnc.Location,
-			RenewDate: tokEnc.RenewDate,
-			Token:     string(tokDec),
-			TeamID:    tokEnc.TeamID,
-			NullTeam:  tokEnc.NullTeam,
-		})
-	}
-
-	return tokens, nil
-}
-
-func (ds *Datastore) GetVPPTokenByTeamID(ctx context.Context, teamID *uint) (*fleet.VPPTokenDB, error) {
-	stmtTeam := `
-	SELECT
-		id,
-		organization_name,
-		location,
-		renew_at,
-		token,
-		team_id,
-		null_team_type
-	FROM
-		vpp_tokens
-	WHERE
-		team_id = ?
-`
-	stmtNullTeam := `
-	SELECT
-		id,
-		organization_name,
-		location,
-		renew_at,
-		token,
-		team_id,
-		null_team_type
-	FROM
-		vpp_tokens
-	WHERE
-		team_id IS NULL
-    AND
-        null_team_type = ?
-`
-
-	var tokEnc struct {
-		ID        uint               `db:"id"`
-		OrgName   string             `db:"organization_name"`
-		Location  string             `db:"location"`
-		RenewDate time.Time          `db:"renew_at"`
-		Token     []byte             `db:"token"`
-		TeamID    *uint              `db:"team_id"`
-		NullTeam  fleet.NullTeamType `db:"null_team_type"`
-	}
-
-	var err error
-	if teamID != nil {
-		err = sqlx.GetContext(ctx, ds.reader(ctx), &tokEnc, stmtTeam, teamID)
-	} else {
-		err = sqlx.GetContext(ctx, ds.reader(ctx), &tokEnc, stmtNullTeam, fleet.NullTeamNoTeam)
-	}
-	if err != nil {
-		if errors.Is(sql.ErrNoRows, err) {
-			if err := sqlx.GetContext(ctx, ds.reader(ctx), &tokEnc, stmtNullTeam, fleet.NullTeamAllTeams); err != nil {
-				if errors.Is(err, sql.ErrNoRows) {
-					return nil, ctxerr.Wrap(ctx, notFound("VPPToken"), "retrieving vpp token by team")
-				}
-				return nil, ctxerr.Wrap(ctx, err, "retrieving vpp token by team")
-			}
-		} else {
-			return nil, ctxerr.Wrap(ctx, err, "retrieving vpp token by team")
-		}
-	}
-
-	tokDec, err := decrypt(tokEnc.Token, ds.serverPrivateKey)
-	if err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "decrypting vpp token with serverPrivateKey")
-	}
-
-	tok := &fleet.VPPTokenDB{
-		ID:        tokEnc.ID,
-		OrgName:   tokEnc.OrgName,
-		Location:  tokEnc.Location,
-		RenewDate: tokEnc.RenewDate,
-		Token:     string(tokDec),
-		TeamID:    tokEnc.TeamID,
-		NullTeam:  tokEnc.NullTeam,
-	}
-
-	return tok, nil
-}
-
-func (ds *Datastore) checkVPPNullTeam(ctx context.Context, currentID *uint, nullTeam fleet.NullTeamType) error {
-	nullTeamStmt := `SELECT id FROM vpp_tokens WHERE null_team_type = ?`
-
-	if nullTeam != fleet.NullTeamNone {
-		var id uint
-		if err := sqlx.GetContext(ctx, ds.reader(ctx), &id, nullTeamStmt, nullTeam); err != nil {
-			if !errors.Is(err, sql.ErrNoRows) {
-				return ctxerr.Wrap(ctx, err, "checking for nullteam constraints")
-			}
-		} else {
-			if currentID == nil || *currentID != id {
-				return ctxerr.Errorf(ctx, "vpp token for team %s already exists", nullTeam)
-			}
-		}
-	}
-
-	return nil
 }
 
 func (ds *Datastore) ListABMTokens(ctx context.Context) ([]*fleet.ABMToken, error) {
